🤖 refactor: Move main.tsx and terminal-window.tsx to browser/
- Moved src/main.tsx → src/browser/main.tsx
- Moved src/terminal-window.tsx → src/browser/terminal-window.tsx
- Updated entry points in index.html and terminal.html
- Fixed relative imports in both files (./browser/api → ./api, etc.)
- Updated tsconfig.main.json exclude list (no longer need explicit excludes)
- Updated eslint.config.mjs to remove separate ignores and update file patterns

These files are React entry points and belong in the browser/ folder.
