+++ index.html
@@ -26,6 +26,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/browser/main.tsx"></script>
   </body>
 </html>
