+++ src/browser/terminal-window.tsx
@@ -8,10 +8,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { TerminalView } from "@/browser/components/TerminalView";
-import "./browser/styles/globals.css";
+import "./styles/globals.css";
 
 // Shims the `window.api` object with the browser API if not running in Electron
-import "./browser/api";
+import "./api";
 
 // Get workspace ID from query parameter
 const params = new URLSearchParams(window.location.search);
