+++ terminal.html
@@ -19,6 +19,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/terminal-window.tsx"></script>
+    <script type="module" src="/src/browser/terminal-window.tsx"></script>
   </body>
 </html>
