+++ tsconfig.main.json
@@ -18,5 +18,5 @@
     "src/node/**/*",
     "src/common/**/*"
   ],
-  "exclude": ["src/browser/**/*", "src/main.tsx", "src/terminal-window.tsx", "src/cli/debug/**/*"]
+  "exclude": ["src/browser/**/*", "src/cli/debug/**/*"]
 }
